Add files via upload
This file contains the functions I have been using to facilitate comparisons. It is currently here to demonstrate how I have been coding my evaluation metric functions, specifically the inputs and outputs, so that Tangram and JSD can be integrated.
Note: the WSIR function calls are not up to date, meaning it currently couldn't be integrated with the function names in the wSIR package. This will be fixed once the recent weighting algorithm becomes the default there.